server/requirement.txt
fastapi
uvicorn
tensorflow==2.12
tensorflow-addons
numpy
pillow
opencv-python
matplotlib
slowapi
limits
starlette
